Some notes:
The way prefix compression works (whereby two leaf nodes may leave off a
substantial portion of their prefix in the path from them to the root)
forces us to store leaves in the table in order to get completion.

This shouldn't be too bad, it will just add to the overall complexity of
the setup.